+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 ## Changelog
 
+### Version 0.17.6-alpha:
+- Piped `.json` files through `jsonesque` first, so comments can exist in the `.json` files.
+- Added an in-programme text-editor.
+- Added an in-programme image-viewer.
+- Fixed opening of files if the value "system" is given.
+- Added a start window.
+- Added a resource pack installer.
+- Added a resource pack opener (opens resource packs into the program, only works with folders).
+- Fixed the zoom-in cap for the image-viewer.
+- Added textures to the buttons on the toolbar of the image-viewer.
+- Added `try`/`except` statements around each loading of an image.
+- Changed `install_pack` to install folders instead of `.zip`s.
+- Added a button to install server resource packs (currently does nothing).
+- Added ToolTips to Buttons on Toolbars.
+- Changed the program to use default editors/viewers if `properties.json` is not found.
+- Changed the image in the image-viewer to scroll with the mouse wheel.
+
 ### Version 0.8.9-alpha:
 - Switched to use `Pillow`, `PIL.Image` and `PIL.ImageTk` classes instead of the `tkinter.PhotoImage` class.
 - Moved images into a folder and switched to use those images instead of `base64` strings.